Add setup.py
This allows the command to be installed in a virtualenv, with auto
dependencies. The command can then be added to the system path and
work anywhere. It can also simplify development and packaging in
the future.

The two gw.py scripts are still available for backwards compatibility.
However, the code from the inner one had to be moved into the package,
so now they are both wrappers. They are not required for anyone using
setup.py.

setup.py will also rebuild version.py automatically.

To install: cd scripts && pip install .

To install in development mode: cd scripts && pip install -e .