Make it work with pyinstaller in docker
Changes to allow pyinstaller to produce a standalone executable. This
does not require any installation. The resulting file is about 6MB on
Windows.

Requires removing the make invocation from setup.py - this will fail
in certain environments if pip copies the source tree somewhere else
before building, as we won't have the Makefile in ..

Requires changing bitarray to bitarray-hardbyte. This is a proof of
concept fork adding wheels support to bitarray, see
https://github.com/ilanschnell/bitarray/issues/63

Before building or installing you must manually generate version.py.

To build for native platform (do this inside a venv to prevent
unnecessary imports):

    cd scripts
    pip install .
    pyinstaller -y --additional-hooks-dir=. --onefile gw.py

To build for Windows using Wine inside Docker on Linux host:

    docker run -v "$(pwd):/src/" cdrx/pyinstaller-windows "pip install . && pyinstaller -y --additional-hooks-dir=. --onefile gw.py"

(Not tested that the resulting exe actually works.)

Cross-building for Mac OS does not seem to be possible, but native
builds should be fine.